% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Finding The Greatest Common Divisor Of Two Numbers}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{one of the number to compute GCD}

\item{y}{another number to compute GCD}
}
\value{
returns HCF of two numbers
}
\description{
It aims to achieve an efficient method for computing the greatest 
common divisor (GCD) of two integers (numbers), the largest number 
that divides them both without a remainder.
}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
