% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{The graph is a data.frame with three variables (v1, v2 and w) that contains 
the edges of the graph (from v1 to v2) with the weight of the edge}
\format{
A graph with 18 rows and 3 variables:
\describe{
  \item{v1}{the edges in the graph}
  \item{v2}{the edges in the graph} 
  \item{w}{the weight between two edges}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra's_algorithm}
}
\usage{
wiki_graph
}
\description{
The graph is a data.frame with three variables (v1, v2 and w) that contains 
the edges of the graph (from v1 to v2) with the weight of the edge
}
\keyword{datasets}
