% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Computes the shortest distance between nodes in graph}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{Inputting the data frame}

\item{init_node}{Imputing the node from which shortest distance must be 
calculated}
}
\value{
shortest distance between nodes
}
\description{
Computes the shortest distance between nodes in graph
}
